Fix some format string issues.
gcc complains about the way we're printing some types like
size_t, ptrdiff_t, etc.  I tried to guarantee that we're printing
them in as portable a way as possible.
Also found an incorrect specifier.

+++ src/fetchmail.c
@@ -852,7 +852,8 @@ static void examineFolder(CURL * handle, struct FOLDER *f, bool dostats)
 
 	nzFree(mailstring);
 	if (dostats) {
-		printf("%2d %s", f - topfolders + 1, f->path);
+		printf("%2lld %s", (unsigned long long)(f - topfolders + 1),
+		       f->path);
 /*
 		if (f->children)
 			printf(" with children");
